feat: Adicionado validação para o endpoint de excluir despesa

+++ src/ControleSeusGastos/Application/Services/Despesas/ExcluirDespesa/ExcluirDespesaService.cs
@@ -1,18 +1,26 @@
-﻿using Infrastructure.Repositories.Depesas;
+﻿using Application.Validacao;
+using Infrastructure.Repositories.Depesas;
 
 namespace Application.Services.Despesas.ExcluirDespesa
 {
     internal class ExcluirDespesaService : IExcluirDespesaService
     {
         private readonly IDespesaRepository _despesaRepository;
+        private readonly IValidadorDatabase _validadorDatabase;
 
-        public ExcluirDespesaService(IDespesaRepository despesaRepository)
+        public ExcluirDespesaService(IDespesaRepository despesaRepository, IValidadorDatabase validadorDatabase)
         {
             _despesaRepository = despesaRepository;
+            _validadorDatabase = validadorDatabase;
         }
 
         public async Task<bool> Excluir(int despesaId)
         {
+            if(await _validadorDatabase.ExisteDespesaDB(despesaId) == false)
+            {
+                return false;
+            }
+
             var afetadas = await _despesaRepository.removerPorId(despesaId);
             if (afetadas > 1)
             {
